refactor(rest): extract routing implementation into standalone files
The file `src/router/routing-table.ts` has grown too much in size and
hosts too many classes, interfaces and helper functions.

This commit is introducing several new files:
 - `src/router/route-entry.ts` to hold
   - RouteEntry
   - ResolvedRoute
   - createResolvedRoute
- `src/router/base-route.ts` for `BaseRoute` class
- `src/router/controller-route.ts` for `ControllerRoute` class
- `src/router/handler-route.ts` for `Route` class
- `src/router/static-assets-route.ts` for `StaticAssetsRoute` class
- `src/router/rest-router.ts` for `RestRouter` class
