+++ packages/rest/src/router/index.ts
@@ -3,9 +3,20 @@
 // This file is licensed under the MIT License.
 // License text available at https://opensource.org/licenses/MIT
 
-export * from './routing-table';
-export * from './openapi-path';
-export * from './trie';
+// routes
+export * from './route-entry';
+export * from './base-route';
+export * from './controller-route';
+export * from './handler-route';
+export * from './static-assets-route';
+
+// routers
+export * from './rest-router';
 export * from './trie-router';
 export * from './regexp-router';
+
+// helpers
+export * from './routing-table';
 export * from './route-sort';
+export * from './openapi-path';
+export * from './trie';
