+++ packages/rest/src/router/route-sort.ts
@@ -1,10 +1,10 @@
-import {RouteEntry} from './routing-table';
-import pathToRegExp = require('path-to-regexp');
-
 // Copyright IBM Corp. 2018. All Rights Reserved.
 // Node module: @loopback/rest
 // This file is licensed under the MIT License.
 // License text available at https://opensource.org/licenses/MIT
+
+import {RouteEntry} from './route-entry';
+import pathToRegExp = require('path-to-regexp');
 
 /**
  * Sorting order for http verbs
